backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
google-genai
python-dotenv